+++ packages/iov-bns/src/bnscodec.ts
@@ -23,7 +23,7 @@ export const bnsCodec: TxCodec = {
   bytesToSign: (tx: UnsignedTransaction, nonce: Nonce): SigningJob => {
     // we encode it without any signatures
     const built = buildUnsignedTx(tx);
-    const bz = codecImpl.app.Tx.encode(built).finish();
+    const bz = codecImpl.bnsd.Tx.encode(built).finish();
     // now we want to append the nonce and chainID
     const bytes = appendSignBytes(bz, tx.creator.chainId, nonce);
     return { bytes: bytes, prehashType: PrehashType.Sha512 };
@@ -32,7 +32,7 @@ export const bnsCodec: TxCodec = {
   // bytesToPost includes the raw transaction appended with the various signatures
   bytesToPost: (tx: SignedTransaction): PostableBytes => {
     const built = buildSignedTx(tx);
-    const out = new Uint8Array(codecImpl.app.Tx.encode(built).finish());
+    const out = new Uint8Array(codecImpl.bnsd.Tx.encode(built).finish());
     return out as PostableBytes;
   },
 
@@ -45,7 +45,7 @@ export const bnsCodec: TxCodec = {
 
   // parseBytes will recover bytes from the blockchain into a format we can use
   parseBytes: (bz: PostableBytes, chainId: ChainId): SignedTransaction => {
-    const parsed = codecImpl.app.Tx.decode(bz);
+    const parsed = codecImpl.bnsd.Tx.decode(bz);
     return parseTx(parsed, chainId);
   },