+++ CHANGELOG.md
@@ -12,6 +12,12 @@ Breaking changes
 
 - @iov/bns: `ElectionRule`, `Proposal` and `VersionedId` all use numeric IDs now
   to conform with `Electorate`.
+- @iov/bns: Remove support for weave 0.16.x; add support for weave 0.17.x.
+- @iov/bns: Remove `AddAddressToUsernameTx` and `RemoveAddressFromUsernameTx` in
+  favour of the new `UpdateTargetsOfUsernameTx`.
+- @iov/bns: Rename `BnsUsernameNft.addresses` to `.targets`.
+- @iov/bns: Rename `RegisterUsernameTx.addresses` to `.targets`.
+- @iov/bns: `RegisterUsernameTx.username` now has to include a \*iov suffix.
 
 ## 0.15.1